Updated gradle from 7.4 to latest (8.11.1) + required build.gradle.kts changes Updated minecraft version from 1.20.4 to 1.21.1. Updated java from 17 to 21. Bumped up version from 4.4.0-BETA to 4.5.0-BETA.
Added damageSource to botDeathEvent as it is now required for 1.21.
Added boolean (false) to comonListenerCookie.createInitial as this is now required.
Updated addEntityPacket method to take all required arguments.
Updated overridden setListener method to setListenerForServerboundHandshake as it seems this is the new method required (needs to be tested).

Updated imports to use org.bukkit.craftbukkit instead of org.bukkit.craftbukkit .v1_20_r3 (it seems craftbukkit updated and versioning is not required anymore).